Implement home page. New home module. Remove set selector module. Styling.
