+++ app/modules/home/genre-list.controller.js
@@ -1,12 +1,13 @@
 'use strict';
 
-require('./set-selector.module.js')
+require('./home.module.js')
     .controller('GenreListController', GenreListController);
 
 // @ngInject
 function GenreListController($state, GenresFactory) {
     var vm = this;
     vm.genres = GenresFactory.list();
+    console.log('Dio cane?')
     vm.go = go;
 
     function go(genre) {
