+++ app/modules/home/genre.constants.js
@@ -1,6 +1,6 @@
 'use strict';
 
-require('./set-selector.module.js')
+require('./home.module.js')
     .constant('GenreListConstants', [
         {
             active: true,
