+++ app/modules/home/home.html
@@ -1,0 +1,42 @@
+<div class="bg-home">
+    <div class="container-fluid">
+        <div class="home-jumbotron">
+
+            <div class="bg"></div>
+            <div class="text">
+                <h3>Folger Shakespeare Library Presents</h3>
+                <div class="separator"></div>
+                <h1>Shakespeare's <b>World</b></h1>
+                <p>Shakespeare's World enables citizen humanists from around the globe to join the effort of transcribing and encoding thousands of early modern manuscripts pages from the Folger Shakespeare Library.</p>
+                <!--
+                <a ui-sref="Transcribe" class="btn btn-success btn-lg">Get Started</a>
+<a ui-sref="GenreList" class="btn btn-success btn-lg">Find Genres</a>
+-->
+            </div>
+            <div class="genre-search">
+                <div class="row">
+                    <div class="content">
+                        <div class="input-group">
+                            <input type="text" ng-model="selected" placeholder="Look up genres" typeahead="genre as genre.name for genre in vm.genres | filter:{name: $viewValue} | limitTo:8" typeahead-on-select="vm.go($item)" typeahead-loading="vm.loading" class="input">
+                            <div class="input-group-addon" aria-hidden="true">
+                                <span class="icon"></span>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
+
+<div class="container">
+    <ul class="genre-list">
+        <h2>Choose a genre and start transcribing!</h2>
+        <li class="genre" ng-repeat="genre in vm.genres" ng-click="vm.go(genre)" ng-class="{ '-active': hover }" ng-mouseenter="hover = true" ng-mouseleave="hover = false">
+            <!--<img ng-src="{{ genre.image }}">-->
+            <div class="title">
+                <h4>{{ genre.name }}</h4>
+            </div>
+        </li>
+    </ul>
+</div>
