+++ app/modules/home/home.module.js
@@ -1,0 +1,5 @@
+'use strict';
+
+// A module for the home page and genres
+module.exports = require('angular')
+    .module('app.home', []);
