+++ app/modules/layout/header/header.directive.js
@@ -24,10 +24,6 @@ function appHeader() {
                 state: 'Home'
             },
             {
-                label: 'Genres',
-                state: 'GenreList'
-            },
-            {
                 label: 'About',
                 state: 'About'
             },
