--- app/modules/set-selector/set-selector.module.js
@@ -1,5 +1,0 @@
-'use strict';
-
-module.exports = require('angular')
-    .module('app.setSelector', [
-    ]);
