--- app/modules/static/home.html
@@ -1,16 +1,0 @@
-<div class="bg-home">
-    <div class="container-fluid">
-        <div class="home-jumbotron">
-
-            <div class="bg"></div>
-            <div class="text">
-                <h3>Folger Shakespeare Library Presents</h3>
-                <div class="separator"></div>
-                <h1>Shakespeare's <b>World</b></h1>
-                <p>Shakespeare's World enables citizen humanists from around the globe to join the effort of transcribing and encoding thousands of early modern manuscripts pages from the Folger Shakespeare Library.</p>
-                <a ui-sref="Transcribe" class="btn btn-success btn-lg">Get Started</a>
-                <a ui-sref="GenreList" class="btn btn-success btn-lg">Find Genres</a>
-            </div>
-        </div>
-    </div>
-</div>
