+++ app/modules/static/static.module.js
@@ -1,5 +1,5 @@
 'use strict';
 
-// A module for the static pages such as home, about etc.
+// A module for the static pages such as about, team, etc.
 module.exports = require('angular')
     .module('app.static', []);
