+++ app/styles/components/bg-home.styl
@@ -2,5 +2,4 @@
     background-image: url("/images/bg-home-sm.jpg")
     background-repeat: no-repeat
     background-size: cover
-    background-attachment: fixed
     background-position: 0% 0%
