+++ app/styles/components/genre-list.styl
@@ -3,7 +3,6 @@
     list-style-type: none
     margin-top: 60px
     padding: 0
-
     > .genre
         @extend .col-xs-6
         @extend .col-sm-4
@@ -14,9 +13,9 @@
         @extend .img-responsive
 
     > .genre > .title
-        background-color: #39403e
+        background-color: offwhite
         border-radius: 50%
-        color: light-grey
+        color: green
         padding: 65px 0
         text-align: center
 
@@ -33,3 +32,7 @@
     > .genre.-active > .title
         background-color: green
         color: #fff
+.genre-list > h2
+    color: offwhite
+    text-align: center
+    margin-bottom: 60px
