+++ app/styles/components/genre-search.styl
@@ -26,11 +26,11 @@
         font-size: 1.5em
 
     > .row > .content > .input-group > .dropdown-menu
-        background-color: orange
+        background-color: green
         border: 0
         border-radius: 0
         color: medium-grey
-        font-size: 20px
+        font-size: 16px
         font-weight: light
         margin: 0
         width: 100%
